refactor: move internal/daemon/ to internal/muxcore/daemon/ (T020)
