+++ cmd/mcp-mux/daemon.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"github.com/thebtf/mcp-mux/internal/muxcore/control"
-	"github.com/thebtf/mcp-mux/internal/daemon"
+	"github.com/thebtf/mcp-mux/internal/muxcore/daemon"
 	"github.com/thebtf/mcp-mux/internal/muxcore/ipc"
 	"github.com/thebtf/mcp-mux/internal/muxcore/serverid"
 )
