+++ internal/muxcore/daemon/snapshot.go
