+++ internal/muxcore/daemon/supervisor_test.go
